backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "markdown>=3.8.2",
    "google-generativeai>=0.8.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-magic>=0.4.27",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
    "debugpy>=1.8.0",
]
